cabal-evals/pyproject.toml
[project]
name = "cabal-evals"
version = "0.1.0"
description = "Production reliability testing for CABAL AI agents"
authors = [
    {name = "CABAL", email = "cabal@openclaw.com"}
]
requires-python = ">=3.10"
dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "pydantic>=2.6.0",
    "anthropic>=0.18.0",
    "python-dateutil>=2.8.2",
]

[project.optional-dependencies]
dev = [
    "black>=24.0.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
